First pass of reduction completed

+++ lib/inc/protocol/scrutiny_protocol_tools.hpp
@@ -132,9 +132,9 @@ namespace scrutiny
             return v;
         }
 
-        uint8_t decode_address_big_endian(uint8_t* buf, uint64_t* addr);
+        uint8_t decode_address_big_endian(uint8_t* buf, uintptr_t* addr);
         uint8_t encode_address_big_endian(uint8_t* buf, void* ptr);
-        uint8_t encode_address_big_endian(uint8_t* buf, uint64_t addr);
+        uint8_t encode_address_big_endian(uint8_t* buf, uintptr_t addr);
     }
 }
