+++ lib/src/protocol/scrutiny_protocol_tools.cpp
@@ -19,36 +19,36 @@ namespace scrutiny
 #pragma warning(push)
 #pragma warning(disable:4127)   // Get rid of constexpr always true condition warning.
 #endif 
-        uint8_t decode_address_big_endian(uint8_t* buf, uint64_t* addr)
+        uint8_t decode_address_big_endian(uint8_t* buf, uintptr_t* addr)
         {
             constexpr unsigned int addr_size = sizeof(void*);
             static_assert(addr_size == 1 || addr_size == 2 || addr_size == 4 || addr_size == 8, "Unsupported address size");
 
-            uint64_t computed_addr = 0;
+            uintptr_t computed_addr = 0;
             unsigned int i = 0;
 
             if (addr_size >= 8)
             {
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 56));
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 48));
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 40));
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 32));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 56));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 48));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 40));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 32));
             }
 
             if (addr_size >= 4)
             {
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 24));
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 16));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 24));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 16));
             }
 
             if (addr_size >= 2)
             {
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 8));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 8));
             }
 
             if (addr_size >= 1)
             {
-                computed_addr |= ((static_cast<uint64_t>(buf[i++]) << 0));
+                computed_addr |= ((static_cast<uintptr_t>(buf[i++]) << 0));
             }
             
             *addr = computed_addr;
@@ -61,7 +61,7 @@ namespace scrutiny
 
         uint8_t encode_address_big_endian(uint8_t* buf, void* ptr)
         {
-            return encode_address_big_endian(buf, reinterpret_cast<uint64_t>(ptr));
+            return encode_address_big_endian(buf, reinterpret_cast<uintptr_t>(ptr));
         }
 
 
@@ -69,7 +69,7 @@ namespace scrutiny
 #pragma warning(push)
 #pragma warning(disable:4127)   // Get rid of constexpr always true condition warning.
 #endif 
-        uint8_t encode_address_big_endian(uint8_t* buf, uint64_t addr)
+        uint8_t encode_address_big_endian(uint8_t* buf, uintptr_t addr)
         {
             constexpr unsigned int addr_size = sizeof(void*);
             static_assert(addr_size == 1 || addr_size == 2 || addr_size == 4 || addr_size == 8, "Unsupported address size");
